+++ src/hash_table/bucket.rs
@@ -32,7 +32,10 @@ pub type DataBlock<K, V, const LEN: usize> = [MaybeUninit<(K, V)>; LEN];
 
 /// [`EntryPtr`] points to an occupied slot in a [`Bucket`].
 pub struct EntryPtr<'b, K: Eq, V, const LOCK_FREE: bool> {
+    /// Points to a [`LinkedBucket`].
     current_link_ptr: Ptr<'b, LinkedBucket<K, V, LINKED_BUCKET_LEN>>,
+
+    /// The index number that the [`EntryPtr`] is pointing to.
     current_index: usize,
 }
 
@@ -46,9 +49,6 @@ pub(crate) struct Metadata<K: Eq, V, const LEN: usize> {
 
     /// Bitmap for removed slots.
     removed_bitmap: u32,
-
-    /// Partial hash array.
-    partial_hash_array: [u8; LEN],
 }
 
 /// [`Locker`] owns a [`Bucket`] by holding the exclusive lock on it.
@@ -100,7 +100,7 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
         &'b self,
         data_block: &'b DataBlock<K, V, BUCKET_LEN>,
         key: &Q,
-        partial_hash: u8,
+        hash: u64,
         barrier: &'b Barrier,
     ) -> Option<&'b (K, V)>
     where
@@ -111,16 +111,14 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
             return None;
         }
 
-        if let Some((_, entry_ref)) =
-            Self::search_entry(&self.metadata, data_block, key, partial_hash)
-        {
+        if let Some((_, entry_ref)) = Self::search_entry(&self.metadata, data_block, key, hash) {
             return Some(entry_ref);
         }
 
         let mut link_ptr = self.metadata.link.load(Acquire, barrier);
         while let Some(link) = link_ptr.as_ref() {
             if let Some((_, entry_ref)) =
-                Self::search_entry(&link.metadata, &link.data_block, key, partial_hash)
+                Self::search_entry(&link.metadata, &link.data_block, key, hash)
             {
                 return Some(entry_ref);
             }
@@ -171,7 +169,7 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
         &self,
         data_block: &DataBlock<K, V, BUCKET_LEN>,
         key: &Q,
-        partial_hash: u8,
+        hash: u64,
         barrier: &'b Barrier,
     ) -> EntryPtr<'b, K, V, LOCK_FREE>
     where
@@ -182,8 +180,7 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
             return EntryPtr::new(barrier);
         }
 
-        if let Some((index, _)) = Self::search_entry(&self.metadata, data_block, key, partial_hash)
-        {
+        if let Some((index, _)) = Self::search_entry(&self.metadata, data_block, key, hash) {
             return EntryPtr {
                 current_link_ptr: Ptr::null(),
                 current_index: index,
@@ -193,7 +190,7 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
         let mut current_link_ptr = self.metadata.link.load(Acquire, barrier);
         while let Some(link) = current_link_ptr.as_ref() {
             if let Some((index, _)) =
-                Self::search_entry(&link.metadata, &link.data_block, key, partial_hash)
+                Self::search_entry(&link.metadata, &link.data_block, key, hash)
             {
                 return EntryPtr {
                     current_link_ptr,
@@ -212,7 +209,7 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
         metadata: &'b Metadata<K, V, LEN>,
         data_block: &'b DataBlock<K, V, LEN>,
         key: &Q,
-        partial_hash: u8,
+        hash: u64,
     ) -> Option<(usize, &'b (K, V))>
     where
         K: Borrow<Q>,
@@ -227,10 +224,8 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
             fence(Acquire);
         }
 
-        let preferred_index = usize::from(partial_hash) % LEN;
-        if (bitmap & (1_u32 << preferred_index)) != 0
-            && metadata.partial_hash_array[preferred_index] == partial_hash
-        {
+        let preferred_index = hash as usize % LEN;
+        if (bitmap & (1_u32 << preferred_index)) != 0 {
             let entry_ref = unsafe { &(*data_block[preferred_index].as_ptr()) };
             if entry_ref.0.borrow() == key {
                 return Some((preferred_index, entry_ref));
@@ -247,11 +242,9 @@ impl<K: Eq, V, const LOCK_FREE: bool> Bucket<K, V, LOCK_FREE> {
         let mut offset = bitmap.trailing_zeros();
         while offset != 32 {
             let index = (preferred_index + offset as usize) % LEN;
-            if metadata.partial_hash_array[index] == partial_hash {
-                let entry_ref = unsafe { &(*data_block[index].as_ptr()) };
-                if entry_ref.0.borrow() == key {
-                    return Some((index, entry_ref));
-                }
+            let entry_ref = unsafe { &(*data_block[index].as_ptr()) };
+            if entry_ref.0.borrow() == key {
+                return Some((index, entry_ref));
             }
             bitmap -= 1_u32 << offset;
             offset = bitmap.trailing_zeros();
@@ -399,19 +392,6 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> EntryPtr<'b, K, V, LOCK_FREE> {
             data_block[self.current_index].as_mut_ptr()
         };
         unsafe { &mut (*entry_ptr) }
-    }
-
-    /// Gets the partial hash value of the entry.
-    ///
-    /// The [`EntryPtr`] must point to an occupied entry.
-    #[inline]
-    pub(crate) fn partial_hash(&self, bucket: &Bucket<K, V, LOCK_FREE>) -> u8 {
-        debug_assert_ne!(self.current_index, usize::MAX);
-        if let Some(link) = self.current_link_ptr.as_ref() {
-            link.metadata.partial_hash_array[self.current_index]
-        } else {
-            bucket.metadata.partial_hash_array[self.current_index]
-        }
     }
 
     /// Tries to remove the [`LinkedBucket`] from the linked list.
@@ -592,14 +572,14 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
         &self,
         data_block: &DataBlock<K, V, BUCKET_LEN>,
         key: &Q,
-        partial_hash: u8,
+        hash: u64,
         barrier: &'b Barrier,
     ) -> EntryPtr<'b, K, V, LOCK_FREE>
     where
         K: Borrow<Q>,
         Q: Eq + ?Sized,
     {
-        self.bucket.get(data_block, key, partial_hash, barrier)
+        self.bucket.get(data_block, key, hash, barrier)
     }
 
     /// Reserves memory for insertion, and then constructs the key-value pair.
@@ -609,18 +589,20 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
     pub(crate) fn insert_with<C: FnOnce() -> (K, V)>(
         &mut self,
         data_block: &mut DataBlock<K, V, BUCKET_LEN>,
-        partial_hash: u8,
+        hash: u64,
         constructor: C,
         barrier: &'b Barrier,
     ) -> EntryPtr<'b, K, V, LOCK_FREE> {
         assert!(self.bucket.num_entries != u32::MAX, "array overflow");
 
+        #[allow(clippy::cast_possible_truncation)]
         let free_index = Self::get_free_index::<BUCKET_LEN>(
             self.bucket.metadata.occupied_bitmap,
-            partial_hash as usize % BUCKET_LEN,
+            hash as usize % BUCKET_LEN,
         );
         if free_index == BUCKET_LEN {
-            let preferred_index = partial_hash as usize % LINKED_BUCKET_LEN;
+            #[allow(clippy::cast_possible_truncation)]
+            let preferred_index = hash as usize % LINKED_BUCKET_LEN;
             let mut link_ptr = self.bucket.metadata.link.load(Acquire, barrier);
             while let Some(link_mut) = unsafe {
                 (link_ptr.as_raw() as *mut LinkedBucket<K, V, LINKED_BUCKET_LEN>).as_mut()
@@ -634,7 +616,6 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
                         &mut link_mut.metadata,
                         &mut link_mut.data_block,
                         free_index,
-                        partial_hash,
                         constructor,
                     );
                     self.bucket.num_entries += 1;
@@ -656,7 +637,6 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
                 link_mut.data_block[preferred_index]
                     .as_mut_ptr()
                     .write(constructor());
-                link_mut.metadata.partial_hash_array[preferred_index] = partial_hash;
                 link_mut.metadata.occupied_bitmap |= 1_u32 << preferred_index;
             }
             if let Some(head) = link.metadata.link.load(Relaxed, barrier).as_ref() {
@@ -679,7 +659,6 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
                 &mut self.bucket.metadata,
                 data_block,
                 free_index,
-                partial_hash,
                 constructor,
             );
             self.bucket.num_entries += 1;
@@ -785,14 +764,12 @@ impl<'b, K: Eq, V, const LOCK_FREE: bool> Locker<'b, K, V, LOCK_FREE> {
         metadata: &mut Metadata<K, V, LEN>,
         data_block: &mut DataBlock<K, V, LEN>,
         index: usize,
-        partial_hash: u8,
         constructor: C,
     ) {
         debug_assert!(index < LEN);
 
         unsafe {
             data_block[index].as_mut_ptr().write(constructor());
-            metadata.partial_hash_array[index] = partial_hash;
             if LOCK_FREE {
                 fence(Release);
             }
@@ -945,7 +922,6 @@ impl<K: Eq, V, const LEN: usize> Default for Metadata<K, V, LEN> {
             link: AtomicArc::default(),
             occupied_bitmap: 0,
             removed_bitmap: 0,
-            partial_hash_array: [0; LEN],
         }
     }
 }
@@ -966,7 +942,6 @@ impl<K: Eq, V, const LEN: usize> LinkedBucket<K, V, LEN> {
                 link: next.map_or_else(AtomicArc::null, AtomicArc::from),
                 occupied_bitmap: 0,
                 removed_bitmap: 0,
-                partial_hash_array: [0; LEN],
             },
             data_block: unsafe { MaybeUninit::uninit().assume_init() },
             prev_link: AtomicPtr::default(),
@@ -1002,7 +977,7 @@ mod test {
 
     use tokio::sync;
 
-    static_assertions::assert_eq_size!(Bucket<String, String, false>, [u8; BUCKET_LEN * 2]);
+    static_assertions::assert_eq_size!(Bucket<String, String, false>, [u8; BUCKET_LEN]);
 
     #[cfg_attr(miri, ignore)]
     #[tokio::test(flavor = "multi_thread", worker_threads = 16)]
@@ -1021,7 +996,7 @@ mod test {
             let data_ptr = AtomicPtr::new(&mut data);
             task_handles.push(tokio::spawn(async move {
                 barrier_clone.wait().await;
-                let partial_hash = (task_id % BUCKET_LEN).try_into().unwrap();
+                let hash = (task_id % BUCKET_LEN).try_into().unwrap();
                 let bucket_mut =
                     unsafe { &mut *(bucket_clone.as_ptr() as *mut Bucket<usize, usize, true>) };
                 let data_block_mut = unsafe {
@@ -1041,14 +1016,14 @@ mod test {
                     if i == 0 {
                         exclusive_locker.insert_with(
                             data_block_mut,
-                            partial_hash,
+                            hash,
                             || (task_id, 0),
                             &barrier,
                         );
                     } else {
                         assert_eq!(
                             exclusive_locker
-                                .search(&data_block_clone, &task_id, partial_hash, &barrier)
+                                .search(&data_block_clone, &task_id, hash, &barrier)
                                 .unwrap(),
                             &(task_id, 0_usize)
                         );
@@ -1058,7 +1033,7 @@ mod test {
                     let read_locker = Reader::lock(&*bucket_clone, &barrier).unwrap();
                     assert_eq!(
                         read_locker
-                            .search(&data_block_clone, &task_id, partial_hash, &barrier)
+                            .search(&data_block_clone, &task_id, hash, &barrier)
                             .unwrap(),
                         &(task_id, 0_usize)
                     );
@@ -1117,7 +1092,7 @@ mod test {
             let data_block_clone = data_block.clone();
             let bucket_clone = bucket.clone();
             task_handles.push(tokio::spawn(async move {
-                let partial_hash = (task_id % BUCKET_LEN).try_into().unwrap();
+                let hash = (task_id % BUCKET_LEN).try_into().unwrap();
                 barrier_clone.wait().await;
                 for _ in 0..256 {
                     loop {
@@ -1140,7 +1115,7 @@ mod test {
                                 let mut exclusive_locker = exclusive_locker.unwrap();
                                 exclusive_locker.insert_with(
                                     data_block_mut,
-                                    partial_hash,
+                                    hash,
                                     || (task_id, 0),
                                     &barrier,
                                 );
@@ -1162,12 +1137,7 @@ mod test {
                                 assert_eq!(
                                     read_locker
                                         .unwrap()
-                                        .search(
-                                            &data_block_clone,
-                                            &task_id,
-                                            partial_hash,
-                                            &barrier,
-                                        )
+                                        .search(&data_block_clone, &task_id, hash, &barrier,)
                                         .unwrap(),
                                     &(task_id, 0_usize)
                                 );
@@ -1186,12 +1156,8 @@ mod test {
                         };
                         let barrier = Barrier::new();
                         let mut exclusive_locker = Locker::lock(bucket_mut, &barrier).unwrap();
-                        let mut entry_ptr = exclusive_locker.get(
-                            &data_block_clone,
-                            &task_id,
-                            partial_hash,
-                            &barrier,
-                        );
+                        let mut entry_ptr =
+                            exclusive_locker.get(&data_block_clone, &task_id, hash, &barrier);
                         assert_eq!(
                             exclusive_locker
                                 .erase(data_block_mut, &mut entry_ptr)